+++ src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/domain/gamereview/service/GameReviewService.kt
@@ -185,9 +185,4 @@ class GameReviewService(
     fun getTopReviews(): List<GameReviewResponse> =
 
         gameReviewRepository.findTopGameReviews().map { it.toResponse() }
-
-
-    fun getGameReviewReactionList(memberId: UUID): List<GameReviewReactionResponse> =
-
-        gameReviewReactionRepository.findByMemberId(memberId).map { it.toResponse() }
 }
